Broke DeferredStore into it's own file.
Also, added stub test file.
